:recycle: refactor(global) : file과 S3 분리
