+++ src/test/java/com/ip/ddangddangddang/domain/file/service/FileServiceTest.java
@@ -14,6 +14,7 @@ import com.ip.ddangddangddang.domain.file.repository.FileRepository;
 import com.ip.ddangddangddang.domain.file.values.FileValues;
 import com.ip.ddangddangddang.domain.user.service.UserService;
 import com.ip.ddangddangddang.global.exception.custom.EmptyImageException;
+import com.ip.ddangddangddang.global.s3.FileUploadService;
 import java.io.IOException;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
